Remove separate MCP server file and integrate FastMCP
