+++ Dockerfile
@@ -3,4 +3,5 @@ WORKDIR /app
 COPY . /app
 RUN pip install uv
 RUN uv pip install -e .[dev]
-CMD ["python", "app.py", "/data/transactions.csv"]
+EXPOSE 8000
+CMD ["uvicorn", "server:app", "--host", "0.0.0.0", "--port", "8000"]
