+++ fhm/__init__.py
@@ -1,0 +1,6 @@
+"""Financial Health Manager core utilities package."""
+
+from .core import parse_csv, savings_rate, summarize
+from .models import Summary
+
+__all__ = ["parse_csv", "savings_rate", "summarize", "Summary"]
