+++ pyproject.toml
@@ -10,7 +10,10 @@ requires-python = ">=3.10"
 # Additional dependencies can be added as needed
 dev = [
     "ruff>=0.0.0",
-    "pytest>=0.0.0"
+    "pytest>=0.0.0",
+    "fastapi>=0.0.0",
+    "uvicorn>=0.0.0",
+    "fastmcp>=0.0.0",
 ]
 
 [tool.ruff]
